+++ src/eda/__init__.py
@@ -2,8 +2,10 @@
 
 from .distribution import DistributionAnalyzer
 from .plots import PlotAnalyzer
+from .tsa import TimeSeriesAnalyzer
 
 __all__ = [
     "PlotAnalyzer",
     "DistributionAnalyzer",
+    "TimeSeriesAnalyzer",
 ]